+++ tests/fixtures/scripts/mapped/fixture_mapped_include_1.lua
@@ -1,0 +1,1 @@
+-- file: fixture_mapped_include_1.lua